Add files via upload
First DDL and ERD commit. 
The SQL file displays the commands that were executed in the process of creating Peak College's database - the entities created and the alteration of those entities to develop a relationship that allows the institution to efficiently organize academic operations, including managing individuals involved in learning, instructional offerings, academic divisions, and participation records. 
The ERD diagrammatically shows the relationship between the entities.